code changes

+++ 0x12-more_singly_linked_lists/1-listint_len.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-#include <stdio.h>
 
 /**
  * listint_len - returns the number of elements in a linked listint_t list.
@@ -9,7 +8,7 @@
  */
 size_t listint_len(const listint_t *h)
 {
-	size_t i;
+	unsigned int i;
 
 	while (h)
 	{
